navdata: add a json() method to each block to get the block as JSON
This patch adds a convenience mthod to the Block classes in navdata.py
which supports returning the black formatted as a JSON message. This may
be parsed in Python into a dictionary via json.loads() from the json
module.

This is a first stage in getting control information *back* to
experimental programs.